Updated dependencies on aiw-i2b2-etl and the parent pom.
Note there is a maven bug that seems to be afflicting compiling eureka-common without first running mvn clean. It appears to be fixed in maven-compiler-plugin version 3.5.1, which is released but has yet to make it out to maven central. We'll want to update the plugin version once we can get a hold of it. In the meantime, you may need to run mvn clean install not mvn install.